requirements.txt
python-dotenv
openai
langchain
langchain-community
langchain-openai
streamlit
brave-search-sdk
PyPDF2
fpdf2
